+++ plugins/plugin-gallery-dl/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Replace inline type definitions with imports from core `plugin-api.d.ts`
+
 ## 2.0.0
 
 ### Breaking Changes